Fix auth resolution for DockerHub.
Fixes a bug where go-sdk was not properly resolving authentication for
DockerHub. For example, pulling a private image from DockerHub would result in
"Error response from daemon: pull access denied for <image-name>, repository does
not exist or may require docker login", even though the user is logged in and
Docker CLI is able to pull the image.

The bug is in function config.resolveAuthConfigForHostname(), where it was
failing to convert/normalize the registry portion of a DockerHub image (e.g.,
docker.io) into something that would match the way the registry is referred to
in the authentication config (i.e., https://index.docker.io/v1). The code
already has a function called auth.ResolveRegistryHostto do this conversion, but
that function was not being called from anywhere. Fix it by calling that
function in config.resolveAuthConfigForHostname().

Signed-off-by: Cesar Talledo <cesar.talledo@docker.com>